During last research level for book, if two guardians are available for current player, the current player has to choose between the 2 guardians.
But any player can overcome freely a guardians.
This is caused by a missing checkaction in the moveToSite function.

Here is a quick fix, that only checks that cuurent player is active.
But the clean fix is probably to move that reseach code out of "digsite" action function.
All special research mechanism would needs such consequent modifications to be cleaner, which is a lot of work and potential unstability.