Fix #39476
When using items :
- fishing rod (19)
- precision compass (20)

Remove the code that was automatically skip the action if not enough coins/compass
Instead, there is a new button "Dont buy item/artefact" if player cannot buy or doesnt want to buy.

If no purshase, the dealt back is put back on the item/artefact deck, as described in the rules.